src/main/java/com/restaurant/service/ReservationService.java
package com.restaurant.service;

import com.restaurant.model.Reservation;

public class ReservationService {

    // Method to create a reservation
    public boolean createReservation(Reservation reservation) {
        // Logic to create a reservation
        // This is a placeholder. Replace with actual implementation.
        return true;
    }

    // Method to update a reservation
    public boolean updateReservation(Reservation reservation) {
        // Logic to update a reservation
        // This is a placeholder. Replace with actual implementation.
        return true;
    }

    // Method to cancel a reservation
    public boolean cancelReservation(int reservationId) {
        // Logic to cancel a reservation
        // This is a placeholder. Replace with actual implementation.
        return true;
    }

    // Method to retrieve a reservation
    public Reservation getReservation(int reservationId) {
        // Logic to get a reservation
        // This is a placeholder. Replace with actual implementation.
        return new Reservation(reservationId, "John Doe", "2024-08-23", "18:00");
    }
}
